requirements.txt
langchain
zhipuai
openai
python-snippets
faiss-cpu